pyproject.toml
[project]
name = "traceloop-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "traceloop-sdk>=0.51.0",
]
